+++ build.gradle
@@ -26,8 +26,6 @@ dependencies {
 	implementation('com.h2database:h2')
 	// test
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-	testImplementation 'org.projectlombok:lombok'
-	testAnnotationProcessor 'org.projectlombok:lombok'
 
 }
