+++ src/main/java/com/delivery/toy/domain/food/dto/request/FindByFoodIdRequest.java
@@ -1,0 +1,8 @@
+package com.delivery.toy.domain.food.dto.request;
+
+import lombok.Builder;
+
+@Builder
+public record FindByFoodIdRequest(
+        Long id) {
+}
