+++ src/main/java/com/delivery/toy/domain/food/exception/FoodIllegalException.java
@@ -1,0 +1,7 @@
+package com.delivery.toy.domain.food.exception;
+
+public class FoodIllegalException extends IllegalArgumentException {
+    public FoodIllegalException(String s) {
+        super(s);
+    }
+}
